Modified Dashbaord

+++ app/routes/api.selected-product.jsx
@@ -3,17 +3,21 @@ import { authenticate } from "../shopify.server";
 import prisma from "../db.server";
 
 export const action = async ({ request }) => {
-  const { admin, shop } = await authenticate.admin(request);
+  const { admin, session } = await authenticate.admin(request);
   const { productIds } = await request.json();
 
+  if (!session?.shop) {
+       return unauthenticated.redirectToAuth({ request });
+  }
+
   await prisma.selectedProduct.deleteMany({
-    where: { shop: shop.shop },
+    where: { shop: session.shop},
   });
 
   await Promise.all(
     productIds.map((id) =>
       prisma.selectedProduct.create({
-        data: { shop: shop.shop, productId: id.toString() },
+        data: { shop: session.shop, productId: id.toString() },
       })
     )
   );
